Address compiler errors about signedness

+++ source/opcode.cpp
@@ -106,7 +106,8 @@ spv_result_t spvOpcodeTableNameLookup(spv_target_env env,
     // Note that the second rule assumes the extension enabling this instruction
     // is indeed requested in the SPIR-V code; checking that should be
     // validator's work.
-    if ((env >= entry.minVersion || entry.numExtensions > 0) &&
+    if ((static_cast<uint32_t>(env) >= entry.minVersion ||
+         entry.numExtensions > 0u) &&
         nameLength == strlen(entry.name) &&
         !strncmp(name, entry.name, nameLength)) {
       // NOTE: Found out Opcode!
@@ -150,7 +151,8 @@ spv_result_t spvOpcodeTableValueLookup(spv_target_env env,
     // Note that the second rule assumes the extension enabling this instruction
     // is indeed requested in the SPIR-V code; checking that should be
     // validator's work.
-    if (env >= it->minVersion || it->numExtensions > 0) {
+    if (static_cast<uint32_t>(env) >= it->minVersion ||
+        it->numExtensions > 0u) {
       *pEntry = it;
       return SPV_SUCCESS;
     }
